Add new command outputting interesting network characteristics

+++ src/pylimer_tools_cpp/entities/UniverseSequence.h
@@ -26,7 +26,7 @@ namespace entities {
     void setDataFileAtomStyle(const std::vector<pylimer_tools::utils::AtomStyle> &dataFileAtomStyle);
 
     // computations
-    std::unordered_map<int, double> computeMsdForAtoms(const std::vector<long int> &atomIds, int nrOfOrigins = 10, bool reduceMemory = false); 
+    std::unordered_map<long int, double> computeMsdForAtoms(const std::vector<long int> &atomIds, int nrOfOrigins = 10, bool reduceMemory = false); 
 
   protected:
     size_t index = 0; // current index of the iterator
@@ -44,8 +44,8 @@ namespace entities {
     Universe readDataFileAtIndex(const size_t index);
     Universe readDumpFileAtIndex(const size_t index);
 
-        std::unordered_map<int, double> computeMsdForAtomsFromDataFiles(const std::vector<long int> &atomIds, int nrOfOrigins = 10, bool reduceMemory = false); 
-        std::unordered_map<int, double> computeMsdForAtomsFromDumpFile(const std::vector<long int> &atomIds, int nrOfOrigins = 10, bool reduceMemory = false); 
+        std::unordered_map<long int, double> computeMsdForAtomsFromDataFiles(const std::vector<long int> &atomIds, int nrOfOrigins = 10, bool reduceMemory = false); 
+        std::unordered_map<long int, double> computeMsdForAtomsFromDumpFile(const std::vector<long int> &atomIds, int nrOfOrigins = 10, bool reduceMemory = false); 
 
   };
 } // namespace entities
